significant service discovery speed up through parallelization + disable blank screen

+++ ModbrosMonitoring/Scripts/stopchrome.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+if [[ $EUID -ne 0 ]]; then
+   echo "This script requires root privileges"
+   exit 1
+fi
+
 pkill -o chromium
 
 exit 0
